+++ src/validate/no-op-validator.ts
@@ -1,4 +1,4 @@
 import { PositiveValidationResult } from './validation-result';
 import { Validator } from './validator';
 
-export const noOpValidator: Validator = () => new PositiveValidationResult();
+export const noOpValidator: Validator = () => PositiveValidationResult.getInstance();